+++ src/sha256.cpp
@@ -172,7 +172,8 @@ constexpr void Step3(const std::size_t t,
                SHA256_sigma_0(W[t - 15]) +  W[t - 16];
     }
 
-    T = (h_) + SHA256_SIGMA_1(e_) + SHA256_Ch(e_, f_, g_) + K_t[t] + W[t];
+    const auto K_t_view = std::span(K_t);
+    T = (h_) + SHA256_SIGMA_1(e_) + SHA256_Ch(e_, f_, g_) + K_t_view[t] + W[t];
 
     d_ += T;
 
@@ -745,16 +746,19 @@ void SHA256::Finalize()
  */
 void SHA256::PadMessage()
 {
+    // Define a span over the input_block array (for bounds checking)
+    auto input_view = std::span(input_block);
+
     // Append 0x80 to the end of the message
-    input_block[input_block_length++] = 0x80;
+    input_view[input_block_length++] = 0x80;
 
     // Pad out to a full input block if we have more than 448 bits (56 octets)
     if (input_block_length > 56)
     {
         // Pad the input block with zeros
-        std::memset(input_block.data() + input_block_length,
-                    0,
-                    64 - input_block_length);
+        std::ranges::fill(input_view.subspan(input_block_length,
+                                             Block_Size - input_block_length),
+                          static_cast<std::uint8_t>(0));
 
         // The input block is now 64 octets, but that will be reset below
 
@@ -768,16 +772,16 @@ void SHA256::PadMessage()
     // Pad up to 448 bits (56 octets)
     if (input_block_length < 56)
     {
-        std::memset(input_block.data() + input_block_length,
-                    0,
-                    56 - input_block_length);
+        std::ranges::fill(
+            input_view.subspan(input_block_length, 56 - input_block_length),
+            static_cast<std::uint8_t>(0));
     }
 
     // The final 64 bits contain the message length (convert length to bits)
     std::uint64_t length = message_length << 3;
     for (std::size_t i = 63; i > 55; i--)
     {
-        input_block[i] = length & 0xff;
+        input_view[i] = length & 0xff;
         length >>= 8;
     }
 
@@ -820,10 +824,13 @@ std::string SHA256::Result() const
 
     oss << std::hex << std::setfill('0');
 
+    // Define a span for the benefit of bounds checking
+    const auto digest_view = std::span(message_digest);
+
     for (std::size_t i = 0; i < Digest_Word_Count; i++)
     {
         if (space_separate_words && (i > 0)) oss << " ";
-        oss << std::setw(8) << message_digest[i];
+        oss << std::setw(8) << digest_view[i];
     }
 
     return oss.str();
@@ -867,12 +874,15 @@ std::span<std::uint8_t> SHA256::Result(std::span<std::uint8_t> result) const
         throw HashException("SHA-256 result input span is too short");
     }
 
+    // Define a span for the benefit of bounds checking
+    const auto digest_view = std::span(message_digest);
+
     for (std::size_t i = 0, j = 0; i < Digest_Word_Count; i++, j += 4)
     {
-        result[j    ] = (message_digest[i] >> 24) & 0xff;
-        result[j + 1] = (message_digest[i] >> 16) & 0xff;
-        result[j + 2] = (message_digest[i] >>  8) & 0xff;
-        result[j + 3] = (message_digest[i]      ) & 0xff;
+        result[j    ] = (digest_view[i] >> 24) & 0xff;
+        result[j + 1] = (digest_view[i] >> 16) & 0xff;
+        result[j + 2] = (digest_view[i] >>  8) & 0xff;
+        result[j + 3] = (digest_view[i]      ) & 0xff;
     }
 
     return result.first(Digest_Octet_Count);